+++ src/pages/index.tsx
@@ -4,14 +4,7 @@ import useDocusaurusContext from "@docusaurus/useDocusaurusContext";
 import Layout from "@theme/Layout";
 import HomepageFeatures from "@site/src/components/HomepageFeatures";
 import Heading from "@theme/Heading";
-
-import {} from "@docusaurus/router";
-
 import styles from "./index.module.css";
-import useMsalAuth from "../hooks/useMsalAuth";
-import { AuthenticatedTemplate, UnauthenticatedTemplate, useAccount, useMsal } from "@azure/msal-react";
-import { Fragment, useEffect, useState } from "react";
-import Login from "../components/ui-components/SignInComponent";
 
 function HomepageHeader() {
   const { siteConfig } = useDocusaurusContext();
@@ -37,55 +30,15 @@ function HomepageHeader() {
 
 export default function Home(): JSX.Element {
   const { siteConfig } = useDocusaurusContext();
-  const { customFields } = useMsalAuth();
-
-  const { accounts } = useMsal();
-  const account = useAccount(accounts[0] || {});
-  const [isAuthorized, setIsAuthorized] = useState(false);
-
-  useEffect(() => {
-    const checkAuthorization = () => {
-      if (!account) return false;
-
-      const userEmail = account.username;
-      const adminDomains = String(customFields.adminDomains).split(",");
-      const externalAccessEmails = String(customFields.externalAccessEmails).split(",");
-      const externalAccessDomains = String(customFields.externalAccessDomains).split(",");
-
-      // Check if the user's email is in the allowed emails list
-      if (externalAccessEmails.includes(userEmail)) return true;
-
-      // Check if the user's email domain is in the allowed domains list
-      const userDomain = userEmail.split("@")[1];
-      if (adminDomains.includes(`@${userDomain}`) || externalAccessDomains.includes(`@${userDomain}`)) return true;
-
-      return false;
-    };
-
-    setIsAuthorized(checkAuthorization());
-  }, [account]);
 
   return (
     <>
-      <Fragment>
-        {customFields.enableAuth == "false" ? (
-          <Layout title={`${siteConfig.title}`} description="Dark Warehouse Docs">
-            <HomepageHeader />
-            <main>
-              <HomepageFeatures />
-            </main>
-          </Layout>
-        ) : isAuthorized ? (
-          <Layout title={`${siteConfig.title}`} description="Dark Warehouse Docs">
-            <HomepageHeader />
-            <main>
-              <HomepageFeatures />
-            </main>
-          </Layout>
-        ) : (
-          <Login />
-        )}
-      </Fragment>
+      <Layout title={`${siteConfig.title}`} description="Dark Warehouse Docs">
+        <HomepageHeader />
+        <main>
+          <HomepageFeatures />
+        </main>
+      </Layout>
     </>
   );
 }
