+++ template.env
@@ -1,0 +1,21 @@
+REACT_APP_ADMIN_DOMAINS="@example-domain.com"
+
+REACT_APP_EXTERNAL_ACCESS_EMAILS="example-email@example-domain.com"
+
+REACT_APP_EXTERNAL_ACCESS_DOMAINS="@example-domain.com"
+
+REACT_APP_MSAL_CLIENT_ID=00000000-0000-0000-0000-000000000000
+
+REACT_APP_MSAL_TENANT_ID=00000000-0000-0000-0000-000000000000
+
+REACT_APP_MSAL_SECRET_ID=00000000-0000-0000-0000-000000000000
+
+REACT_APP_MSAL_SECRET_VALUE=example-secret-key
+
+REACT_APP_MSAL_REDIRECT_URI_DEVELOPMENT=http://localhost:3000/
+
+REACT_APP_MSAL_REDIRECT_URI_PRODUCTION=https://docs.darkwarehouse.io
+
+REACT_APP_ENABLE_AUTH=true
+
+REACT_APP_TEST_LOCALLY=true
